1282 refactor generatePdfFromHtml to useCaseHelper
