+++ shared/src/business/useCases/generatePdfFromHtmlInteractor.js
@@ -1,3 +1,4 @@
+import { generatePdfFromHtml } from '../useCaseHelper/pdf/generatePdfFromHtml';
 /**
  * generatePdfFromHtmlInteractor
  *
@@ -15,66 +16,11 @@ exports.generatePdfFromHtmlInteractor = async ({
   docketNumber,
   headerHtml,
 }) => {
-  let browser = null;
-  let result = null;
-
-  try {
-    applicationContext.logger.time('Generating Docket Record PDF');
-    const chromium = applicationContext.getChromium();
-
-    browser = await chromium.puppeteer.launch({
-      args: chromium.args,
-      defaultViewport: chromium.defaultViewport,
-      executablePath: await chromium.executablePath,
-      headless: true,
-    });
-
-    let page = await browser.newPage();
-
-    await page.setContent(contentHtml);
-
-    const headerTemplate = `
-      <!doctype html>
-      <html>
-        <head>
-        </head>
-        <body style="margin: 0px;">
-          <div style="font-size: 8px; font-family: sans-serif; width: 100%; margin: 0px 40px; margin-top: 25px;">
-            <div style="font-size: 8px; font-family: sans-serif; float: right;">
-              Page <span class="pageNumber"></span>
-              of <span class="totalPages"></span>
-            </div>
-            <div style="float: left">
-              ${headerHtml ? headerHtml : `Docket Number: ${docketNumber}`}
-            </div>
-          </div>
-        </body>
-      </html>
-    `;
-
-    const footerTemplate = `
-      <div style="font-size:8px !important; color:#000; text-align:center; width:100%; margin-bottom:5px;">Printed <span class="date"></span></div>
-    `;
-
-    result = await page.pdf({
-      displayHeaderFooter,
-      footerTemplate: footerTemplate,
-      format: 'Letter',
-      headerTemplate: headerTemplate,
-      margin: {
-        bottom: '100px',
-        top: '80px',
-      },
-      printBackground: true,
-    });
-  } catch (error) {
-    applicationContext.logger.error(error);
-    throw error;
-  } finally {
-    if (browser !== null) {
-      await browser.close();
-    }
-  }
-  applicationContext.logger.timeEnd('Generating Docket Record PDF');
-  return result;
+  return generatePdfFromHtml({
+    applicationContext,
+    contentHtml,
+    displayHeaderFooter,
+    docketNumber,
+    headerHtml,
+  });
 };
